+++ src/main/java/g6shenpcare/repository/CustomerProfileRepository.java
@@ -3,7 +3,7 @@ package g6shenpcare.repository;
 import g6shenpcare.entity.CustomerProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerProfileRepository extends JpaRepository<CustomerProfile, Long> {
+public interface CustomerProfileRepository extends JpaRepository<CustomerProfile, Integer> {
 
     long countByIsActiveTrue();
